fixup! Add discard message feedback toast
